apps/api/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# GraphQL
strawberry-graphql[fastapi]==0.235.0

# Database
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache
redis[hiredis]==5.0.7

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AI
anthropic==0.31.0
instructor==1.3.3

# Config & validation
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1

# HTTP client
httpx==0.27.0

# Background jobs
celery[redis]==5.4.0

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
